backend/requirements.txt
pytest
pandas
numpy
matplotlib
seaborn
flask
flask-cors
azure-storage-blob
python-dotenv
